+++ src/main/kotlin/de/nordgedanken/auto_hotkey/runconfig/core/AhkRunConfigType.kt
@@ -1,20 +1,18 @@
 package de.nordgedanken.auto_hotkey.runconfig.core
 
 import com.intellij.execution.configurations.ConfigurationFactory
-import com.intellij.execution.configurations.ConfigurationType
 import com.intellij.execution.configurations.ConfigurationTypeBase
 import com.intellij.execution.configurations.ConfigurationTypeUtil
 import de.nordgedanken.auto_hotkey.util.AhkConstants
 import de.nordgedanken.auto_hotkey.util.AhkIcons
-import org.jetbrains.annotations.Nls
 
 /**
  * Creates a new type of run config to select when choosing a run config template. Registered in plugin.xml
  */
 class AhkRunConfigType : ConfigurationTypeBase(
-    AhkConstants.RunConfig.ID,
+    ID,
     AhkConstants.LANGUAGE_NAME,
-    AhkConstants.RunConfig.DESCRIPTION,
+    DESCRIPTION,
     AhkIcons.LOGO
 ) {
     init {
@@ -28,6 +26,9 @@ class AhkRunConfigType : ConfigurationTypeBase(
     }
 
     companion object {
+        const val ID = "AhkRunConfiguration"
+        const val DESCRIPTION = "AutoHotkey run configuration"
+
         fun getInstance(): AhkRunConfigType =
             ConfigurationTypeUtil.findConfigurationType(AhkRunConfigType::class.java)
     }