+++ src/main/kotlin/de/nordgedanken/auto_hotkey/util/AhkConstants.kt
@@ -4,9 +4,4 @@ object AhkConstants {
     const val PLUGIN_NAME = "AutoHotkey_Language"
     const val LANGUAGE_NAME = "AutoHotkey"
     const val FILE_EXTENSION = "ahk"
-
-    object RunConfig {
-        const val ID = "AhkRunConfiguration"
-        const val DESCRIPTION = "AutoHotkey run configuration"
-    }
 }